+++ src/resources/extensions/shared/format-utils.ts
@@ -1,11 +1,11 @@
 /**
- * Shared formatting and layout utilities for TUI dashboard components.
+ * Shared pure formatting utilities — no @gsd/pi-tui dependency.
  *
- * Consolidates helpers that were previously duplicated across
- * auto-dashboard.ts, dashboard-overlay.ts, and visualizer-views.ts.
+ * ANSI-aware layout helpers (padRight, joinColumns, centerLine, fitColumns)
+ * live in layout-utils.ts to avoid pulling @gsd/pi-tui into modules that
+ * run outside jiti's alias resolution (e.g. HTML report generation via
+ * dynamic import in auto-loop).
  */
-
-import { truncateToWidth, visibleWidth } from "@gsd/pi-tui";
 
 // ─── Duration Formatting ──────────────────────────────────────────────────────
 
@@ -29,45 +29,6 @@ export function formatTokenCount(count: number): string {
   if (count < 1000) return `${count}`;
   if (count < 1_000_000) return `${(count / 1000).toFixed(1)}k`;
   return `${(count / 1_000_000).toFixed(2)}M`;
-}
-
-// ─── Layout Helpers ───────────────────────────────────────────────────────────
-
-/** Pad a string with trailing spaces to fill `width` (ANSI-aware). */
-export function padRight(content: string, width: number): string {
-  const vis = visibleWidth(content);
-  return content + " ".repeat(Math.max(0, width - vis));
-}
-
-/** Build a line with left-aligned and right-aligned content. */
-export function joinColumns(left: string, right: string, width: number): string {
-  const leftW = visibleWidth(left);
-  const rightW = visibleWidth(right);
-  if (leftW + rightW + 2 > width) {
-    return truncateToWidth(`${left}  ${right}`, width);
-  }
-  return left + " ".repeat(width - leftW - rightW) + right;
-}
-
-/** Center content within `width` (ANSI-aware). */
-export function centerLine(content: string, width: number): string {
-  const vis = visibleWidth(content);
-  if (vis >= width) return truncateToWidth(content, width);
-  const leftPad = Math.floor((width - vis) / 2);
-  return " ".repeat(leftPad) + content;
-}
-
-/** Join as many parts as fit within `width`, separated by `separator`. */
-export function fitColumns(parts: string[], width: number, separator = "  "): string {
-  const filtered = parts.filter(Boolean);
-  if (filtered.length === 0) return "";
-  let result = filtered[0];
-  for (let i = 1; i < filtered.length; i++) {
-    const candidate = `${result}${separator}${filtered[i]}`;
-    if (visibleWidth(candidate) > width) break;
-    result = candidate;
-  }
-  return truncateToWidth(result, width);
 }
 
 // ─── Text Truncation ─────────────────────────────────────────────────────────
